Add tests for event UDID + tidy up existing tests
